requirements.txt
bilibili_api==3.1.3
Pillow==8.0.1
qrcode==6.1
quart==0.13.1
